src/test/java/spencercjh/problems/RemoveMaxNumberOfEdgesToKeepGraphFullyTraversableTest.java
package spencercjh.problems;

import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
import org.junit.jupiter.api.Test;

import javax.inject.Inject;

import static org.junit.jupiter.api.Assertions.assertEquals;

@MicronautTest
class RemoveMaxNumberOfEdgesToKeepGraphFullyTraversableTest {

  @Inject
  private RemoveMaxNumberOfEdgesToKeepGraphFullyTraversable solution;

  @Test
  void maxNumEdgesToRemove() {
    assertEquals(2, solution.maxNumEdgesToRemove(4, new int[][]{{3, 1, 2}, {3, 2, 3}, {1, 1, 3}, {1, 2, 4}, {1, 1, 2}, {2, 3, 4}}));
    assertEquals(0, solution.maxNumEdgesToRemove(4, new int[][]{{3, 1, 2}, {3, 2, 3}, {1, 1, 4}, {2, 1, 4}}));
  }

}
